+++ aws-java-sdk-s3/src/main/java/com/amazonaws/services/s3/transfer/internal/CopyMonitor.java
@@ -16,14 +16,6 @@ package com.amazonaws.services.s3.transfer.internal;
 
 import static com.amazonaws.event.SDKProgressPublisher.publishProgress;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.Callable;
-import java.util.concurrent.CancellationException;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Future;
-import java.util.concurrent.atomic.AtomicReference;
-
 import com.amazonaws.SdkClientException;
 import com.amazonaws.event.ProgressEventType;
 import com.amazonaws.event.ProgressListenerChain;
@@ -33,6 +25,13 @@ import com.amazonaws.services.s3.model.PartETag;
 import com.amazonaws.services.s3.transfer.Transfer.TransferState;
 import com.amazonaws.services.s3.transfer.TransferManager;
 import com.amazonaws.services.s3.transfer.model.CopyResult;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.Callable;
+import java.util.concurrent.CancellationException;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Future;
+import java.util.concurrent.atomic.AtomicReference;
 
 /**
  * Monitors an copy operation by periodically checking to see if the operation is
@@ -155,7 +154,7 @@ public class CopyMonitor implements Callable<CopyResult>, TransferMonitor {
         markAllDone();
         transfer.setState(TransferState.Completed);
         // Since the copy has completed we can assume all bytes were successfully transferred
-        // This is required since there are no progress updates available during server side
+        // This is required since there are no progress updates available during server-side
         // copying of data.
         transfer.getProgress().updateProgress(transfer.getProgress().getTotalBytesToTransfer());
         // AmazonS3Client takes care of all the events for single part uploads,